modularization WIP
